+++ cmd/gozenviamock/main.go
@@ -1,0 +1,22 @@
+package main
+
+import (
+	"context"
+	"flag"
+	"os"
+
+	"github.com/jeffersonsc/gozenviamock/cmd"
+
+	"github.com/google/subcommands"
+)
+
+func main() {
+	subcommands.Register(subcommands.HelpCommand(), "")
+	subcommands.Register(subcommands.FlagsCommand(), "")
+	subcommands.Register(subcommands.CommandsCommand(), "")
+	subcommands.Register(&cmd.Server{}, "")
+
+	flag.Parse()
+	ctx := context.Background()
+	os.Exit(int(subcommands.Execute(ctx)))
+}
