+++ go.mod
@@ -2,4 +2,10 @@ module github.com/jeffersonsc/gozenviamock
 
 go 1.12
 
-require github.com/gorilla/mux v1.7.3
+require (
+	github.com/google/subcommands v1.0.1
+	github.com/gorilla/handlers v1.4.2
+	github.com/gorilla/mux v1.7.3
+	github.com/sirupsen/logrus v1.4.2 // indirect
+	gocloud.dev v0.16.0 // indirect
+)
